Fix some warnings that only triggered in LINT.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@136862 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/contrib/dev/hptmv/hptproc.c
@@ -275,7 +275,7 @@ static int hpt_set_info(int length)
         			piop->lpInBuffer,
         			piop->nInBufferSize,
         			piop->lpOutBuffer,
-	        		piop->nOutBufferSize));
+	        		(u_long)piop->nOutBufferSize));
 
 			/*
         	 	 * map buffer to kernel.
